Split #1 for psx
